+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ## UNRELEASED
 
+- Fixed: Correctly handle packages, such as itertools, that have a transitive dependency on themselves,
+  by passing a fully-qualified URL to `cargo --package`.
+
 - Don't mutate `unsafe` fns.
 
 - Don't mutate functions that never return (i.e. `-> !`).